requirements.txt
anyascii==0.1.7
asgiref==3.3.0
beautifulsoup4==4.8.2
boto3==1.16.22
botocore==1.19.22
certifi==2020.6.20
chardet==3.0.4
Django==3.1.3
django-extensions==3.0.9
django-filter==2.4.0
django-modelcluster==5.1
django-storages==1.10.1
django-taggit==1.3.0
django-treebeard==4.3.1
djangorestframework==3.12.1
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
html5lib==1.1
idna==2.10
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
l18n==2020.6.1
MarkupSafe==1.1.1
mysqlclient==2.0.1
openpyxl==3.0.5
Pillow==8.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tablib==2.0.0
Unidecode==1.1.1
urllib3==1.25.11
wagtail==2.11
wagtail-font-awesome-svg==0.0.2
webencodings==0.5.1
Willow==1.4
xlrd==1.2.0
XlsxWriter==1.3.7
xlwt==1.3.0
